Moved thumbnail components to their own dir
